Move VUU token parsing into vuu-auth
Relocate parseVuuUserFromToken and the token's VuuUser contract to the authentication package.
