+++ vuu-ui/packages/vuu-data-remote/src/index.ts
@@ -1,7 +1,6 @@
 export {
   authenticate,
   getVuuAuthToken,
-  parseVuuUserFromToken,
 } from "./auth/authenticate";
 export * from "./ConnectionManager";
 export { default as ConnectionManager } from "./ConnectionManager";
@@ -9,7 +8,10 @@ export * from "./constants";
 export * from "./data-source";
 export { LostConnectionHandler, RetryOptions } from "./LostConnectionHandler";
 export * from "./message-utils";
-export { VuuAuthenticator, VuuAuthTokenIssuePolicy } from "./auth/VuuAuthenticator";
+export {
+  VuuAuthenticator,
+  VuuAuthTokenIssuePolicy,
+} from "./auth/VuuAuthenticator";
 export { VuuAuthProvider } from "./auth/VuuAuthProvider";
 export * from "./VuuDataSource";
 export { isConnected, type ConnectionStatus } from "./WebSocketConnection";
